+++ TVMonitorUI/TVMonitorUI.csproj
@@ -6,4 +6,8 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\IMDB_API_Accessor\IMDB_API_Accessor.csproj" />
+  </ItemGroup>
+
 </Project>
